# medienverwaltungweb.client/src/app/api-client/model/actor.ts
/**
 * medienVerwaltungWeb.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Actor { 
    firstName?: string | null;
    id?: number;
    lastName?: string | null;
    movieIds?: Array<number> | null;
    birthDate?: string;
    gender?: string | null;
    readonly fullName?: string | null;
}