# medienverwaltungweb.client/swagger.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "medienVerwaltungWeb.Server",
      "version": "1.0"
    },
    "paths": {
      "/api/Actor": {
        "get": {
          "tags": [
            "Actor"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Actor"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Actor"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Actor"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Actor"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Actor/{id}": {
        "get": {
          "tags": [
            "Actor"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Actor"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Actor"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Actor"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/api/Book": {
        "get": {
          "tags": [
            "Book"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDTO"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDTO"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDTO"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Book"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Book/{isbn}": {
        "get": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Actor": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "lastName": {
              "type": "string",
              "nullable": true
            },
            "movieIds": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "nullable": true
            },
            "birthDate": {
              "type": "string",
              "format": "date-time"
            },
            "gender": {
              "type": "string",
              "nullable": true
            },
            "fullName": {
              "type": "string",
              "nullable": true,
              "readOnly": true
            }
          },
          "additionalProperties": false
        },
        "BookDTO": {
          "type": "object",
          "properties": {
            "isbn": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "interpretFullName": {
              "type": "string",
              "nullable": true
            },
            "pageCount": {
              "type": "integer",
              "format": "int32"
            },
            "releaseYear": {
              "type": "integer",
              "format": "int32"
            },
            "location": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }